+++ src/models/__init__.py
@@ -1,5 +1,6 @@
 from src.models.order import Order, OrderStatus
 from src.models.order_history import OrderStatusHistory
+from src.models.order_image import OrderImage
 from src.models.user import Role, User
 
 __all__ = [
@@ -8,4 +9,5 @@ __all__ = [
     "Order",
     "OrderStatus",
     "OrderStatusHistory",
+    "OrderImage",
 ]